Move `models` into `crates_io_database` crate
... and some `utils` for now until we can separate them properly 😅
