+++ crates/crates_io_database/Cargo.toml
@@ -8,10 +8,28 @@ edition = "2021"
 workspace = true
 
 [dependencies]
-diesel = "=2.2.7"
+bon = "=3.3.2"
+chrono = { version = "=0.4.39", default-features = false, features = ["serde"] }
+crates_io_diesel_helpers = { path = "../crates_io_diesel_helpers" }
+crates_io_index = { path = "../crates_io_index" }
+diesel = { version = "=2.2.7", features = ["serde_json", "chrono", "numeric"] }
+diesel-async = "=0.5.2"
 diesel_full_text_search = "=2.2.0"
+futures-util = "=0.3.31"
+rand = "=0.9.0"
+secrecy = "=0.10.3"
+semver = { version = "=1.0.25", features = ["serde"] }
+serde = { version = "=1.0.217", features = ["derive"] }
+serde_json = "=1.0.138"
+sha2 = "=0.10.8"
+thiserror = "=2.0.11"
+tracing = "=0.1.41"
+unicode-xid = "=0.2.6"
 
 [dev-dependencies]
+claims = "=0.8.0"
 crates_io_test_db = { path = "../crates_io_test_db" }
 diesel-async = { version = "=0.5.2", features = ["postgres"] }
+googletest = "=0.13.0"
+insta = "=1.42.1"
 tokio = { version = "=1.43.0", features = ["macros", "rt"] }
