+++ crates/crates_io_database/src/lib.rs
@@ -1,3 +1,5 @@
 #![doc = include_str!("../README.md")]
 
+pub mod models;
 pub mod schema;
+pub mod utils;
