+++ crates/crates_io_database/src/models/deleted_crate.rs
@@ -1,9 +1,9 @@
+use crate::schema::deleted_crates;
 use bon::Builder;
 use chrono::{DateTime, Utc};
-use crates_io_database::schema::deleted_crates;
 
 /// Struct used to `INSERT` a new `deleted_crates` record into the database.
-#[derive(Insertable, Debug, Builder)]
+#[derive(diesel::Insertable, Debug, Builder)]
 #[diesel(table_name = deleted_crates, check_for_backend(diesel::pg::Pg))]
 pub struct NewDeletedCrate<'a> {
     #[builder(start_fn)]
