+++ crates/crates_io_database/src/models/dependency.rs
@@ -1,9 +1,9 @@
-use diesel::sql_types::{BigInt, Text};
-
 use crate::models::{Crate, Version};
 use crate::schema::*;
 use crates_io_diesel_helpers::pg_enum;
 use crates_io_index::DependencyKind as IndexDependencyKind;
+use diesel::prelude::*;
+use diesel::sql_types::{BigInt, Text};
 
 #[derive(Identifiable, Associations, Debug, Queryable, QueryableByName, Selectable)]
 #[diesel(
