+++ crates/crates_io_database/src/models/helpers.rs
