+++ src/util.rs
@@ -1,11 +1,10 @@
 pub use self::io_util::{read_fill, read_le_u32};
 pub use self::request_helpers::*;
+pub use crates_io_database::utils::{rfc3339, token};
 
 pub mod diesel;
 pub mod errors;
 mod io_util;
 mod request_helpers;
-pub mod rfc3339;
 pub mod string_excl_null;
-pub mod token;
 pub mod tracing;
